Loan_Agent/requirements.txt
psycopg[binary]>=3.2.0
python-dotenv>=1.0.1
openai>=1.12.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

